LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointer-events-none-skip-scroll-in-iframe-expected.txt



FAIL Wheel-scroll over pointer-events: none scroller skips that scroller assert_equals: Incorrect element scrolled expected "root" but got "scroller"
